requirements.txt
load_dotenv
streamlit
PyPDF2
langchain
altair
faiss-cpu
tiktoken
huggingface-hub
sentence-transformers
langchain_community
SpeechRecognition
pyaudio
langchain_huggingface
ipykernel
langchain_groq
googletrans==4.0.0-rc1
langdetect
groq 
